Move server tests
