--- grpc-api/src/test/java/io/deephaven/grpc_api/runner/DeephavenApiServerSingleUnauthenticatedBase.java
@@ -1,16 +1,0 @@
-package io.deephaven.grpc_api.runner;
-
-import io.deephaven.grpc_api.DeephavenChannel;
-import org.junit.Before;
-
-public abstract class DeephavenApiServerSingleUnauthenticatedBase extends DeephavenApiServerTestBase {
-
-    DeephavenChannel channel;
-
-    @Override
-    @Before
-    public void setUp() throws Exception {
-        super.setUp();
-        channel = channel();
-    }
-}
