+++ server/src/test/app.d/00-static.app
